requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
httpx==0.28.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pydantic==2.10.2
